# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MindTrace",
    "description": "Service for managing mindtrace",
    "version": "3.1.3"
  },
  "paths": {
...
    "/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Current User",
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Невалидный или истёкший access-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.invalid_access_token",
                  "message": "Невалидный или истёкший access-токен",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "users.user_not_found",
                  "message": "Пользователь не найден",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "410": {
            "description": "Пользователь удалён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "users.user_deleted",
                  "message": "Пользователь удалён",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "internal",
                  "message": "Внутренняя ошибка сервера",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CurrentUserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname",
            "description": "Отображаемое имя; null, если пользователь его не задал."
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "displayName"
        ],
        "title": "CurrentUserResponse"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Код ошибки для программной обработки",
            "examples": [
              "invalid_input",
              "not_found",
              "internal"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Человекочитаемое сообщение об ошибке",
            "examples": [
              "Некорректный запрос",
              "Ресурс не найден"
            ]
          },
          "details": {
            "$ref": "#/components/schemas/OptionalDict",
            "description": "Дополнительные детали ошибки (опционально)",
            "examples": [
              {
                "field": "email",
                "reason": "Invalid email format"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Временная метка возникновения ошибки (UTC)"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Базовая схема ответа ошибки API.\n\nИспользуется для всех типов ошибок и автоматически генерирует документацию OpenAPI.",
        "example": {
          "code": "invalid_input",
          "message": "Некорректный запрос",
          "timestamp": "2024-01-01T12:00:00Z"
        }
      },
...
      "OptionalDict": {
        "anyOf": [
          {
            "additionalProperties": true,
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}